refactor(whatsrust): extract read synchronization wrappers
